Feat: 녹화 종료 시 사용자 연습 시간 업데이트

+++ BackEnd/src/main/java/com/ssafy/ploud/domain/user/UserEntity.java
@@ -45,8 +45,8 @@ public class UserEntity {
   @NotNull
   private LocalDateTime joinDate;
 
-  private int soloDurationInMinute;
-  private int studyDurationInMinute;
+  private long soloDurationInMinute;
+  private long studyDurationInMinute;
 
   @Enumerated(EnumType.STRING)
   private Role role;
@@ -85,11 +85,11 @@ public class UserEntity {
     this.password = password;
   }
 
-  public void updateSoloDuration(int practiceTimeInMinute) {
+  public void updateSoloDuration(long practiceTimeInMinute) {
     this.soloDurationInMinute += practiceTimeInMinute;
   }
 
-  public void updateStudyDuration(int practiceTimeinMinute) {
+  public void updateStudyDuration(long practiceTimeinMinute) {
     this.studyDurationInMinute += practiceTimeinMinute;
   }
